pyproject.toml
[tool.poetry]
name = "dre-repo"
version = "0.2.0"
description = ""
authors = ["DRE Team <dept-DRE@dfinity.org>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10.0,<4"
ansible = ">=7.0.0,<10.0.0"
colorama = "*"
elasticsearch = "<8.0,>=7.16"
elasticsearch-dsl = "*"
elasticsearch-follow = "*"
GitPython = "*"
numpy = "*"
pandas = "*"
paramiko = "*"
PyYAML = "*"
requests = "*"
tabulate = "*"
peewee = "*"
atlassian-python-api = "*"
python-dotenv = "*"
cachetools = "*"
pre-commit = "*"
pylint = "*"
tenacity = "*"
poetry = "^1.6.1"
ic-py = "*"
mkdocs-material = "^9.5.9"
mkdocs = "^1.5.3"
clickhouse-connect = "^0.7.0"
humanfriendly = "^10.0"
jupyter = "^1.0.0"
matplotlib = "^3.8.2"
ipython = "^8.21.0"
mkdocs-git-revision-date-localized-plugin = "^1.2.4"
mkdocs-git-committers-plugin-2 = "^2.2.3"
pydiscourse = "^1.6.1"
datamodel-code-generator = "^0.25.4"
pydantic-yaml = "^1.2.1"
google-api-python-client = "^2.119.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.0"
pydrive2 = "^1.19.0"
markdownify = "^0.11.6"
pytest = "^8.0.2"
pygithub = "^2.2.0"
pytest-mock = "^3.12.0"

[tool.poetry.group.dev.dependencies]
black = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
